actualizando proyecto

+++ src/main/java/com/backendtuscuentas/controller/UsuarioController.java
@@ -29,8 +29,12 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.multipart.MultipartFile;
 
+import com.backendtuscuentas.entitys.Departamento;
 import com.backendtuscuentas.entitys.LogActividade;
 import com.backendtuscuentas.entitys.Menu;
+import com.backendtuscuentas.entitys.Municipio;
+import com.backendtuscuentas.entitys.TipoDocumento;
+import com.backendtuscuentas.entitys.TipoUsuario;
 import com.backendtuscuentas.entitys.Usuario;
 import com.backendtuscuentas.entitys.util.Alerta;
 import com.backendtuscuentas.entitys.util.ListadoEjecuciones;
@@ -73,17 +77,19 @@ public class UsuarioController {
 			String nombreAnterior = usuario.getUsuaImgperfil();
 
 			System.out.println("---------->>> " + usuario.getUsuaImgperfil());
+			// si es la primera vez no se elimina nada
+			if (!nombreAnterior.equals("profile.png")) {
+				if (nombreAnterior != null && nombreAnterior.length() > 0) {
+					Path rutaFotoAnterior = Paths.get("C:\\JavaCode\\archivosTuscuentas\\imagenesperfil")
+							.resolve(nombreAnterior).toAbsolutePath();
 
-			if (nombreAnterior != null && nombreAnterior.length() > 0) {
-				Path rutaFotoAnterior = Paths.get("C:\\JavaCode\\archivosTuscuentas\\imagenesperfil")
-						.resolve(nombreAnterior).toAbsolutePath();
+					File archivoFotoAnterior = rutaFotoAnterior.toFile();
 
-				File archivoFotoAnterior = rutaFotoAnterior.toFile();
+					if (archivoFotoAnterior.exists() && archivoFotoAnterior.canRead()) {
+						archivoFotoAnterior.delete();
+					}
 
-				if (archivoFotoAnterior.exists() && archivoFotoAnterior.canRead()) {
-					archivoFotoAnterior.delete();
 				}
-
 			}
 
 			Path rutaArchivo = Paths.get("C:\\JavaCode\\archivosTuscuentas\\imagenesperfil").resolve(nombreArchivo)
@@ -109,6 +115,10 @@ public class UsuarioController {
 
 	@GetMapping("/upload/img/{nombreFoto:.+}")
 	public ResponseEntity<Resource> verFoto(@PathVariable String nombreFoto) {
+
+		if (nombreFoto.equals("profile.png")) {
+			nombreFoto = "perfil.png";
+		}
 
 		Path rutaArchivo = Paths.get("C:\\JavaCode\\archivosTuscuentas\\imagenesperfil").resolve(nombreFoto)
 				.toAbsolutePath();
@@ -136,38 +146,66 @@ public class UsuarioController {
 
 	@GetMapping("/listarAlertas/{id}")
 	public ResponseEntity<ArrayList<Alerta>> listarAlertasPorIdUsuario(@PathVariable Long id) {
-		return new ResponseEntity<ArrayList<Alerta>>(this.usuarioService.listarAlertasPorIdUsuario(id),
-				HttpStatus.OK);
+		return new ResponseEntity<ArrayList<Alerta>>(this.usuarioService.listarAlertasPorIdUsuario(id), HttpStatus.OK);
 	}
-	
-	
+
 	@GetMapping("/listarAbiertas/{id}/{idEmpresa}")
-	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesAbiertas(@PathVariable Long id,@PathVariable Long idEmpresa) {
-		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(this.usuarioService.listarEjecucionesAbiertas(id, idEmpresa),
-				HttpStatus.OK);
+	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesAbiertas(@PathVariable Long id,
+			@PathVariable Long idEmpresa) {
+		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(
+				this.usuarioService.listarEjecucionesAbiertas(id, idEmpresa), HttpStatus.OK);
 	}
-	
+
 	@GetMapping("/listarAtrasadas/{id}/{idEmpresa}")
-	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesAtrasadas(@PathVariable Long id,@PathVariable Long idEmpresa) {
-		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(this.usuarioService.listarEjecucionesAtrasadas(id, idEmpresa),
-				HttpStatus.OK);
+	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesAtrasadas(@PathVariable Long id,
+			@PathVariable Long idEmpresa) {
+		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(
+				this.usuarioService.listarEjecucionesAtrasadas(id, idEmpresa), HttpStatus.OK);
 	}
-	
+
 	@GetMapping("/listarObservadas/{id}/{idEmpresa}")
-	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesObservadas(@PathVariable Long id,@PathVariable Long idEmpresa) {
-		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(this.usuarioService.listarEjecucionesObservadas(id, idEmpresa),
-				HttpStatus.OK);
+	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesObservadas(@PathVariable Long id,
+			@PathVariable Long idEmpresa) {
+		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(
+				this.usuarioService.listarEjecucionesObservadas(id, idEmpresa), HttpStatus.OK);
 	}
-	
+
 	@GetMapping("/listarCerradas/{id}/{idEmpresa}")
-	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesCerradas(@PathVariable Long id,@PathVariable Long idEmpresa) {
-		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(this.usuarioService.listarEjecucionesCerradas(id, idEmpresa),
+	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesCerradas(@PathVariable Long id,
+			@PathVariable Long idEmpresa) {
+		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(
+				this.usuarioService.listarEjecucionesCerradas(id, idEmpresa), HttpStatus.OK);
+	}
+
+	@GetMapping("/listarProximas/{id}/{idEmpresa}")
+	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesProximas(@PathVariable Long id,
+			@PathVariable Long idEmpresa) {
+		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(
+				this.usuarioService.listarEjecucionesProximas(id, idEmpresa), HttpStatus.OK);
+	}
+
+	@GetMapping(value = "/listarTipoDocumento")
+	public ResponseEntity<List<TipoDocumento>> listarTipoDocumento() {
+		List<TipoDocumento> us = (List<TipoDocumento>) usuarioService.listarTipoDocumento();
+		return new ResponseEntity<List<TipoDocumento>>(usuarioService.listarTipoDocumento(), HttpStatus.OK);
+	}
+
+	@GetMapping(value = "/listarTipoUsuario/{tipo}")
+	public ResponseEntity<List<TipoUsuario>> listarTipoUsuario(@PathVariable Long tipo) {
+		List<TipoUsuario> us = (List<TipoUsuario>) usuarioService.listarTipoUsuario(tipo);
+		return new ResponseEntity<List<TipoUsuario>>(usuarioService.listarTipoUsuario(tipo), HttpStatus.OK);
+	}
+
+	@GetMapping(value = "/listarMunicipios/{idDepartamento}")
+	public ResponseEntity<List<Municipio>> listarMunicipiosPorDepartamento(@PathVariable Long idDepartamento) {
+		List<Municipio> us = (List<Municipio>) usuarioService.listarMunicipioPorDepartamento(idDepartamento);
+		return new ResponseEntity<List<Municipio>>(usuarioService.listarMunicipioPorDepartamento(idDepartamento),
 				HttpStatus.OK);
 	}
-	
-	@GetMapping("/listarProximas/{id}/{idEmpresa}")
-	public ResponseEntity<ArrayList<ListadoEjecuciones>> listarEjecucionesProximas(@PathVariable Long id,@PathVariable Long idEmpresa) {
-		return new ResponseEntity<ArrayList<ListadoEjecuciones>>(this.usuarioService.listarEjecucionesProximas(id, idEmpresa),
-				HttpStatus.OK);
+
+	@GetMapping(value = "/listarDepartamentos")
+	public ResponseEntity<List<Departamento>> listarDepartamentos() {
+		List<Departamento> us = (List<Departamento>) usuarioService.listarDepartamentos();
+		return new ResponseEntity<List<Departamento>>(usuarioService.listarDepartamentos(), HttpStatus.OK);
 	}
 }
